ya multiplica y suma

+++ app/src/main/java/com/example/imvcol/FrmInventarios.java
@@ -44,7 +44,7 @@ public class FrmInventarios extends AppCompatActivity {
 
                 for (int i = 0; i < inventarios.length; i++) {
                     LstItem item = new LstItem(inventarios[i][0] + "-" + (inventarios[i][1] == null ? "Sin nombre" : inventarios[i][1]),
-                            Integer.parseInt(inventarios[i][2].toString()),
+                            Double.parseDouble(inventarios[i][2].toString()),
                             inventarios[i][2 + usuario.getCurrConteo()] == null ? "Falta" : inventarios[i][2 + usuario.getCurrConteo()].toString());
 
                     data.add(item);
