+++ app/src/main/java/com/example/imvcol/Item/LstItem.java
@@ -2,10 +2,10 @@ package com.example.imvcol.Item;
 
 public class LstItem {
     private String nombre;
-    private int cantidad;
+    private Double cantidad;
     private String conteo;
 
-    public LstItem(String nombre, int cantidad, String conteo) {
+    public LstItem(String nombre, Double cantidad, String conteo) {
         this.nombre = nombre;
         this.cantidad = cantidad;
         this.conteo = conteo;
@@ -19,11 +19,11 @@ public class LstItem {
         this.nombre = nombre;
     }
 
-    public int getCantidad() {
+    public Double getCantidad() {
         return cantidad;
     }
 
-    public void setCantidad(int cantidad) {
+    public void setCantidad(Double cantidad) {
         this.cantidad = cantidad;
     }
